Add API to forward exception to Http2MultiplexHandler streams Motivation:
Some exceptions on the parent channel are of interest to individual HTTP2 streams as well and should be forwarded. This behavior is already hardcoded for SSLException, but cannot be customized.

Modification:

Add a method to Http2MultiplexHandler to fire an exception to the child handlers.

Result:

A downstream handler on the parent channel can intercept exceptions and forward them to the stream channels using the new API.

Not sure if this is the best approach. Maybe `forEachActiveStream` should be exposed somehow. Another alternative would be to allow forwarding user events, and letting the API user wrap/unwrap exceptions in such user events where necessary.