Avoid failing cookie validation on missing space after ';'
Motivation:
In practice, this appears to be a common deviation from the standard.
It's unlikely that we will be able to change all the worlds clients and servers to strictly follow the spec, so we should instead accept this by default.
If we do not, we will probably instead just see people disable cookie validation entirely, along with a lot of support requests.

Modification:
Add a pedantic cookie validation flag, and have pedantic validation disabled by default.
Also add the relevant tests.

Result:
We tolerate more cookies that people will see in practice, without compromising the rest of the cookie validation.