Fix buffer leaks in the codec-http2 module
Motivation:
Manually enabling leak detection for codec-http2, reveals buffer leaks
in tests and production code.

Modification:
- Ensure the original data is closed by CompressorHttp2ConnectionEncoder#writeData
as the ownership is not transferred to Compressor.compress
- Ensure the copied data is closed by DefaultHttp2ConnectionDecoder#onGoAwayRead0
as the ownership is not transferred to Http2FrameListener.onGoAwayRead
- Ensure the created buffer is closed by DelegatingDecompressorFrameListener#onDataRead
as the ownership is not transferred to Http2FrameListener.onDataRead
- Http2FrameCodec.FrameListener#onUnknownFrame should not copy
the DefaultHttp2UnknownFrame as the buffer was splitted
- Fix leaks in tests

Result:
No more leaks in codec-http2.