Speed up HTTP/2 header iteration
Motivation:
When iterating HTTP/2 headers, the pseudo-headers must always come first in iteration order.
We previously solved this by iterating the headers twice, filtering, and concatenating the two filtered iterations.
This was very slow, but meant we could use the MultiMap pretty much unchanged.

Modification:
Add another pair of linkage references to the entries in the MultiMap.
This doubly-linked list preserves total insertion order, except pseudo-headers are always arranged before normal headers.
We use a handful of fields to keep track of where pseudo-headers start and end, and likewise for normal headers.
For iteration we then use this new chain instead of the bucket-head and entries chains.

Result:
Iteration of HTTP/2 headers is now over 700% faster, but 'add' and 'addAll' methods are slower by 22-44%, when benchmarked with 30 headers.