Make the EmbeddedEventLoop thread-safe
Motivation:
Event loops and executors are expected to be thread-safe, and accessible from multiple threads.
Even though the EmbeddedEventLoop is _usually_ only used by one thread, it might not always be.
Especially as we're increasingly running our tests in parallel.

Modification:
Make all of the state-mutating methods of EmbeddedEventLoop synchronized.
The EmbeddedEventLoop extends the AbstractScheduledEventExecutor, which assumes that only one thread will be mutating its state - the event loop thread.
This is not the case with the EmbeddedEventLoop, as it has no dedicated thread but instead relies on external inputs driving its execution.
By making all state mutating methods synchronized, we not only protect the internals from concurrent modification, but also ensure safe publication of the internal state when it is accessed by multiple threads in a non-concurrent way.
Also fix an issue in Http2FrameCodecTest where we didn't wait for a future to complete before checking if it had succeeded or not.

Result:
The Http2FrameCodecTest, which was suffering data races in the AbstractScheduledEventExecutor and supporting classes, now runs more stably.