Make EmbeddedEventLoop detect concurrent access
Motivation:
The EmbeddedEventLoop is meant to only be access by one thread at a time.

Modification:

Add concurrent access checks to the EmbeddedEventLoop, a bit like ConcurrentModificationException in the Java collections.
The EmbeddedEventLoop and EmbeddedChannel are meant to be used by only one thread at any given time.
By having these checks, we are more likely to identify test isolation failures involving the embedded channels.

Result:
We are now much more likely to get a meaningful exception thrown when an EmbeddedEventLoop is accessed by multiple threads.