Fix FlowControlHandler's autoreading breakage. (#12467)
Motivation:

FlowControlHandler breaks auto-reading (i.e. the channel does not auto-read even though its config has auto-read set to true). This happens when auto-read is turned on when the queue is not empty. The reason is that FlowControlHandler.read() does not pass the call to the ChannelHandlerContext if at least one message has been dequeued. This is a problem when FlowControlHandler.read() has been called as a result of auto-read having been turned on (DefaultChannelConfig.setAutoRead() calling read() in case of autoRead && !oldAutoRead).

Modifications:

Always pass the call to read() down the pipeline.

Result:

Auto-reading works, regardless of the handler's queue state in the moment it is turned on.
Fixes #12445