Fix scalability issue of ByteToMessageDecoder sub-classes that also implement ChannelOutboundHandler
Motivation:

There is some scalability issue with ChannelHandler implementations that implement ChannelInboundHandler and ChannelOutboundHandler at the same time. This can be fixed by usually extending ChannelDuplexHandler. Unfortunally this is not possible when the ChannelHandler in question already extend ByteToMessageDecoder, which is the case for SslHandler which is a very popular handler. To fix this let's just have ByteToMessageDecoder extend ChannelDuplexHandler. This will not hurt in general as we will still skip execution of the outbound methods if te sub-class is not overriding these while also ensure that we not get into trouble of scalability issues.

Modifications:

Let ByteToMessageDecoder extend ChannelDuplexHandler

Result:

Workaround scalability issue of SslHandler.