Optimize Hpack encoding for headers with many values (#12665)
Motivation:
The current implementation of the hpack dynamic table hashes the header name, but not the header value, in order to assign a bucket to a header. When a header name is associated with many values, all the (name, value) combinations hash into the same bucket. This degrades the performance of searching in this hash bucket.

Changes:
Use 2 tables, one keyed by header name and the other by header name + value.
the table keyed by name + value hashes both the name and the value in order to assign a bucket to the header.
Also, to lower the table's load factor, the default number of buckets was increased from 16 to 64.
This was done since the default table size (4096) can easily result in more than 64 headers populating the table, which gives a load factor > 4 for 16 buckets, which is much higher than e.g. Java's HashMap default load factor of 0.75.

Results:
Faster encoding for the case of headers with many values, no significant regression for the case of header names not mapping to multiple values.

Co-authored-by: ahadadi <ahadadi@outbrain.com>