Switch test logging to Log4j2
Motivation:
This allows us to use their async appenders, which moves the IO to a background thread, and allow our tests, which run in parallel, to not be impeded by IO and console-stream synchronisation.

Modification:
Remove the logback test dependencies and add log4j2 dependencies instead.
Also add a test dependency on LMAX disruptor, since that is required for the async appenders.
The LoggingHandlerTest and SingleThreadEventLoopTest had compile-time dependencies on logback, and they have been updated to use Log4j2 instead.

Result:
Our builds should be a bit faster now.