Turn ReadHandle.estimatedBufferCapacity into prepareRead
Motivation:
For asynchronous transport implementations we need to know a priori how many messages we can receive in one event-loop iteration for a channel.
To do this we need a slight change to how ReadHandle works.

Modification:
Change the ReadHandle.estimatedBufferCapacity method into a prepareRead method.
The prepareRead may change state of the handle, and may limit the number of buffers that can be prepared at any one time.
For instance, handles derived from the MaxMessagesReadHandleFactory will refuse to prepare more than the max-messages number of buffers.

Result:
Asynchronous transports can now prepare the correct number of read operations, following the instructions of the ReadHandle.