Transport to close channel if read detects closure
Motivation:
Protocols powered by the ByteStream transport classes (TCP, UDS) do not have
independent state for input/output shutdown. If we read the transport is shut
down then future writes are not expected to succeed, but today we leave the
channel in an unusable state and expect the next write to detect the failure
and trigger a close.

Modifications:
- Instead of waiting for the write to detect the channel is closed the transport
  should just close the channel when it detects it is closed.