Make it possible to complete and compact bytes in the `IovecArray`
Motivation:
When we experience incomplete reads or writes, it can be useful to compact and reuse the same `IovecArray` that the first IO was submitted with.
Like a smaller, more local version of what the `OutboundBuffer` is doing.
This makes it easier to drive certain system calls to "completion" in some cases.

Modification:
Remove the unused `writtenMessages` method, and put a `completeBytes` method in its place.
The `completeBytes` method update length and address of partially completed iovecs, and move the remaining incomplete iovecs to the start of the array, allowing more to be added if need be.

Result:
The `IovecArray` is easier to reuse when faced with incomplete IO.