Don't unnecessarily reorder http headers (#12921)
Motivation:

The guard against multiple Content-Length headers would unconditionally
replace the Content-Length header(s), this means that the Content-Length
header would always be placed last. Changing the order of headers can
have significant impact on code that for example does cryptographic
signing of the headers and should be avoided if possible.

Modification:

Before re-assigning the Content-Length header, check if the current
header is good enough and avoid the (re-)assigning in that case.

Result:

If there is only a single Content-Length header, it will remain in the
same order among the headers.