Fix bug in partially consuming ChannelOutboundBuffer (#12947)
Motivation:
If the callback to ChannelOutboundBuffer.consumeEachFlushedMessage returned `false` to halt the iteration early, the message that caused the halt will then be considered to have been consumed.
That's a bug. The message should instead remain flushed in the outbound buffer.

Modification:
Fix the loop logic so we avoid removing any message that causes the loop to break early.

Result:
The callbackt to ChannelOutboundBuffer.consumeEachFlushedMessage can now break the loop early, if it can't process the given message, without loosing any data.