Attach lifecycle traces to Send.receive exceptions (#12953)
Motivation:
A Send can only be received once. Any subsequent call to receive will throw an exception.
It will be helpful for debugging such exceptions, if they include a lifecycle trace for the received object, which may include the point where the object was first received.

Modification:
Make SendFromOwned remember the received object and, if not null (e.g. via a race), use it to attach traces to any gate exceptions it throws.

Result:
Exceptions from Send.receive are now easier to debug.