Use the recommended directory layout for native-image metadata. Simplify automatic handler metadata generator usage. (#12951)
Motivation:

The recommended directory layout for native-image metadata is `META-INF/native-image/<groupId>/<artifactId>/...`. Recent Native Build Tools versions will print a warning if a `native-image.properties` file is detected and it's path doesn't adhere to the above layout.

Modification:

This PR moves all native-image related metadata so that it conforms to the above layout.
This PR also ensures the above layout is used for the automatically generated handler metadata. The new layout of this metadata is `META-INF/native-image/<groupId>/<artifactId>/automatic/handlers`.

Result:

Fixes #12934