Make it possible for AbstractChannel subclasses to drive the write-loop
Motivation:
For asynchronous transports, we need to be able to drive the write-loop on our own schedule.

Modification:
Pull the write-loop out of `WriteSink` into a protected method on `AbstractChannel`, and offer methods to step and end the write-loop.

Result:
Asynchronous transports can now drive the write-loop forward in response to IO completions.