Provide specialized writeBytes override in NioBuffer
Motivation:

The NioBuffer implementation does not override the default Buffer
version of writeBytes (array and ByteBuffer) leading to reduced
performance for large arrays or buffers due to how the base methods are
implemented, such as arrays which get written a byte at a time
triggering a bounds check on each setByte call.

Modifications:

Implement writeBytes overrides in NioBuffer to increase performance of
code that writes arrays or ByteBuffers into a NioBuffer

Result:

Improved throughput in application that utilize these methods.