Deprecate ObjectEncoder/ObjectDecoder
Motivation:
Serialization has proven itself to be a security liability, and one that cannot easily be fixed: new gadgets are created and found regularly, as are filter-bypasses.

Modification:
Deprecate all of `io.netty.handler.codec.serialization` to discourage its use.

Result:
Use of object serialization is now discouraged.

Thank you to Oriya Yavnieli of the JFrog Security Research team for bringing this to our attention.