Fix SocketHalfClosedTest
Motivation:
This test was not running for epoll, and disabled for io_uring.

Modification:
Get the tests running for epoll and io_uring again.
Fix handling of write errors in epoll and io_uring.
Also update the test to handle interleaving reads and writes, which is something that can happen with io_uring - in contrast, epoll completes all of its writes in one go before doing any reads, hence the write-error and shutdown-output happens before any reads for epoll.

Result:
SocketHalfClosedTest not runs and pass for both epoll and io_uring.