Fix data race bug in UnorderedThreadPoolEventExecutorTest (#12986)
Motivation:
This test could sometimes (about 1 in 100.000) fail because addListener and promise completion could race to notify listeners, causing an extra task to be observed by the queue inspection, after the CountDownLatch had opened.

Modification:
Use an Exchanger to delay the completion of the first task until after we've attached a listener to the promise of the second task.
This way, the completion of the second task will no longer race with `addListener`.
This works because the UnorderedThreadPoolEventExecutor used by the test is single-threaded.

Result:
The UnorderedThreadPoolEventExecutorTest is no longer flaky.