HPACK dynamic table size update must happen at the beginning of the h… (#12988)
Motivation:

HpackDecoder accepts dynamic table size updates at any stage of a header block. Such modification should always happen before the table is modified as per RFC 7541.

Modifications:

Check the header block first for dynamic table size updates in a separate method. Then reject any table size update when processing unpacking the rest of the block. Dynamic table size update could have been done with the current switch/case but it requires to keep track of the modifications which is not hard but introduces complexity in the loop (a simple boolean that must be switched for any modification but the table size update). Instead a table size update is achieved first in a simpler loop and then any table size update is rejected.

Result:

Fixes #12981

Co-authored-by: Norman Maurer <norman_maurer@apple.com>