Use Stream Compression instead of Single-shot Compression (#12924)
Motivation:
As per the current implementation, `BrotliEncoder` compress a `ByteBuf` in a single operation. This approach is good when we have small data set to work with. However, In the case of stream, there are multiple `ByteBuf`, and compressing each `ByteBuf` using the Single-shot method is the wrong approach.

Modification:
Use streaming compression.

Result:
Fixes #12909.