Add MemorySegment Buffer implementation
Motivation:
To ease our maintenance burden, we'd like to consolidate our repositories when the code in question is considered "core" to Netty.
Implementations of the Buffer API certainly qualifies.
We may also, in the future, be able to follow the path that Lucene has taken, and offer MemorySegment support through multi-release jars, but this PR only lays the groundwork for building with a snapshot Java 20, and moves the code over.

Modification:
Add a docker compose build that prepares a snapshot version of Panama JDK 20.
Add a module with the memory segment buffer implementation, and enable it in Maven when we detect Java 20 in the build.

Result:
We're able to run our tests on Java 20 (snapshot, this will be simpler once Java 20 is actually released), and we can test our memory segment buffer implementation with it.