Make TrafficShapingHandlerTest isolated to reduce impact on timers
It's not easy to see, but the handler and/or the test is relying timeliness for asserting correctness properties.
Start and stop times are measured with `TrafficCounter.milliSecondFromNano()`, and a certain amount of traffic needs to be processed in a set time, or else the test fails.
This might not be possible when tests run concurrently, as the overloaded system might be too slow.