Add `FUNDING.yml` (#13012)
We now have a registered non-profit entity in South Korea. We could start to accept donation from the community to cover the cost related with running the projects under @netty GitHub organization, such as (but not limited to):

- Build and dev infra, SaaS subscriptions and equipment
- Hosting and attending the community events
- Paying the core developers for their work; and
- Establishing a non-profit tax-deductible foundation.

We're currently a non-profit entity located in South Korea (this is where @trustin lives). Please note that the donation is not tax-deductible at the moment.