Make testSendToUnresolvableAddress run faster
Motivation:
This test was taking 2 minutes to create the unresolved inet address.

Modification:
By changing to an unresolvable sub-domain of a real address, we can discover that the address cannot be resolved much faster compared to an unresolvable top-level domain.

Result:
Locally, the test now takes ~5 seconds instead of 2 minutes.