Add the AdaptivePoolingAllocator
Motivation:
The current PoolingBufferAllocator has a number of issues.
The code has proven complex to maintain over the years, and has caused a steady trickle of bugs despite its age.
It also makes use of ThreadLocals, which is problematic when used with VirtualThreads as there might be millions of such threads, making the memory usage of PoolThreadLocalCache prohibitively expensive.

Modification:
Instead of using ThreadLocals, the AdaptivePoolingAllocator reduces contention by striping access across a number of magazines.
Each thread will have a preferred magazine to access, computed from their thread ID.
We use StampedLocks to tryLock the magazine, and if the lock is not immediately available we will continue on to the next magazine.
If there is too much contention (too many tryLocks that fail) then the number of magazines is increased, up to twice the number of CPU cores.
This behavior is similar in spirit to Stripe64, though the implementation is quite different.

Each magazine does pointer-bump (using Buffer.split) allocations out of a chunk.
The default chunk size is 128 KiB.
However, allocation sizes are continuously recorded in a sliding-window histogram.
Every 10.000 allocations to a magazine, the 99-percentile allocation size is computed from the histogram, and then multiplied by 10.
If this value is greater than 128 KiB, then it will be used as the new preferred chunk size.
From then on, chunks that are less than this, or over 50% larger than this, will be retired and the memory will be freed.

The maximum size we can track in the histograms is 1 MiB, and the minimum size is 64 bytes.
The sizes are bucketed into powers of two.
The sliding window also only keep 4 generations of histograms, and we use shorts for bucket counters since we rotate them every 10.000 allocations.
This keeps our memory spent on histograms low.
This also means our max chunk size is 10 MiB.

The allocator has a central queue where magazines can pluck available chunks.
However, magazines also have a nextInLine field, which will be checked first - if it is empty we will instead take a chunk from the central queue, or allocate a new one if the queue was empty.
When all of the buffers allocated from a chunk have been closed, then the chunk returns to the allocator.
When this happens we first check if the nextInLine field of the originating magazine is null - if it is, we atomically return the chunk to that magazine.
Otherwise the chunk is given to the central queue.

The dynamic adjustments we do to the chunk sizes and the number of magazines, in response to the workload, is why this allocator is called adaptive.

Result:
We have a new pooling allocator with many benefits:

* It does not rely on thread locals.
* It uses less memory that the existing pooling allocator in the practical applications that have been observed so far.
* It is inherently robust to fragmentation as long as buffers are eventually closed.
* It is faster and (perhaps surprisingly) appear to scale better to multi-threaded workloads than the existing allocator.

These are the benchmark results I get from BufferAllocatorConcurrentBenchmark on an M1 Pro machine:

Benchmark                (size)   Mode  Cnt       Score      Error   Units
allocateReleaseAdaptive   00064  thrpt   20  134000.741 ± 2052.369  ops/ms
allocateReleasePooled     00064  thrpt   20   82131.491 ± 1149.161  ops/ms
allocateReleaseUnpooled   00064  thrpt   20      97.436 ±  109.054  ops/ms