Extend DNS address supplier interface to provide feedback (#12991)
Motivation:

This change allows the start of building out a resolver that is more performant. Most standalone DNS resolvers have some mechanism to find and prioritize nameservers that are giving the best service (either lower error rate or faster response times)

E.g. https://unbound.docs.nlnetlabs.nl/en/latest/reference/history/info-timeout-server-selection.html

Modification:

Added a new interface that extends DnsServerAddressStream to provide a mechanism for DnsResolverContext to provide feedback back to the class that is providing DNS servers in the form of InetSocketAddress

Result:

The DnsServerAddressStream is now able to know the response times of queries for each InetSocketAddress that it returned, or that the InetSocketAddress had a failure when used

Co-authored-by: Ryan Watson <rfwatson@apple.com>