Support handshake timeout in SniHandler. (#13041)
Motivation:
The SslHandler has a configurable timeout that is triggered when the handshake is not completed within that limit. The SniHandler should also have a similar timeout, such timeout would be fired when the the client has not sent enough data to trigger the SNI completion and propagated to the SslHandler it creates.

Modifications:
Added a handshake timeout in the SniHandler that creates a timer task to fire a failed sni completion event, as well as setting the handshake timeout on the SslHandler when it creates it.

Result:
Consequently, SniHandler supports handshake timeouts.

Co-authored-by: Norman Maurer <norman_maurer@apple.com>