Optimize method: PoolChunk#allocateRun(int)
Motivation:

when removeAvailRun(queue, handle) is called within the allocateRun(int), the handle is already removed from queue. So, no need to iterate through entire queue to remove the handle.

Modification:

Moved the logic for removing the handle from the queue to a more appropriate location.

Result:

Removed unnecessary iteration.