Allow to adjust the GlobalEventExecutor quietPeriod via a system property (#13141)
Motivation:

The GlobalEventExecutor will shutdown its "internal thread" after the used quietPeriod if there is no more work to do. This quietPeriod was hardcoded to 1 second which might not be long enough for some applications and so will cause a lot of threads be started and destroyed.

Modifications:

Add io.netty.globalEventExecutor.quietPeriodSeconds property which allows the user to configure the quietPeriod. Default is still 1 second

Result:

Be able to influence the quietPeriod of GlobalEventExecutor via a system property