Only do extra CNAME query if we couldnt follow the whole CNAME chain in the response
Motivation:

If we receive a response with a CNAME chain in it we should only do an extra query if we could follow the whole chain. Otherwise we will produce more queries if necessary and so might overload the remote DNS server.

Modifications:

- Only do extra CNAME query if we couldnt follow the whole chain.
- Add unit test

Result:

Produce less queries when handling CNAMEs