Add unit tests that ensure we correctly handle partial timeouts when doing DNS queries
Motivation:

When we do an A and AAAA query and one is successful while the other times out we need to ensure we still not fail the promise. This was already the case but we didnt have tests for this.

Modifications:

Added unit test to validate correct behaviour

Result:

More testing around timeout handling