Fix FlowControlHandler's behaviour to pass read events when auto-reading is turned off (#13146)
Motivation:
Currently, in FlowControlHandler read events are passed onto the pipeline in every case regardless of the auto-reading configuration of the channel.

Modification:
The read method in FlowControlHandler has been adjusted accordingly.

Result:
Not passing read events onto the pipeline (calling ctx.read() which results in reading from the client socket) when the queue is not empty and auto-reading is turned off for the channel.

Fixes #13104.

Co-authored-by: Angelov <ivan.angelov01@sap.com>