Create a deep-copy of the Throwable before returning it from the cache to prevent possible leaks (#13160)
Motivation:

As we dont know what exactly the user will do with the cached exception we should better ensure to return a copy every time the user tries to access it. The problem is that the exception might stay in the cache for a very long time while the user might call things like addSuppressed(...) on it. This could lead to the sitation that we hold up a lot of references and so "leak" memory until the entry expires.

Modifications:

- Do a deep copy of the cached Throwable before returning it to the user.
- Add unit tests

Result:

No possible to "leak memory"