Fix Host->:authority translation when URI is in origin-form (#213)
Motivation:

Before this patch, HTTP/1->HTTP/3 conversion would only add the :authority pseudo header if the URI was in absolute- or authority-form. However, when there is a Host header, there is no reason why it couldn't be translated in asterisk-form or origin-form as well.

Modification:

If a Host header is present, translate it to the :authority pseudo-header, for all request URI types.

Result:

Translation produces the :authority header for asterisk-form and origin-form request URIs, when the Host header is present.