Consistent maxOrder value
Motivation:

The current implementation of the maxOrder value in the Netty5 is inconsistent, as a recent change (https://github.com/netty/netty/pull/12108) reduced the value from 11 to 9, but this change was not well propagated to Netty 5.

Modification:

Updating the maxOrder value in Netty 5 to match the consistent value of 9.

Result:

Improve consistency. Eliminate unexpected behavior.