Don't add results to finalResults once promise was notified (#13196)
Motivation:

We need to ensure we release the results if we dont end up transfering the ownership. In some cases we could have ended up to adding more results to the List without be sure that these will ever be released at all.

Modifications:

Only add results as long as we did not notify yet. Otherwise release these.

Result:

No more memory leak possible even if promise is already completed